Insert a random number of pseudo cell addresses
For general text files with = prefixed to each line to force
the parser to try and interpret it as an expression where cell
addresses could be there, I found that my internal logic based
parser is 2 times as fast as the re based parser logic.

Just to verify that both parser are handling the parsing of
cell addresses properly, now I am adding some cell addresses
randomly into the lines.