Insert checks for existance of fwdLinks
If there is no fwdLinks associated with a cell, then it means that the
cell doesnt contain any =expression with reference to cell addresses
in it. So whatever was the old calculated value, the same can be
copied into the new calc cache.

However if there was fwdLinks, then it means the cell contains a
=expression, which may require adjusting due to insert operation.

In this case, the old calculated value is copied into the new calc
cache, only if the cell address adjustments didnt raise/tag any errors.

	Because insert on its own doesnt change the value of any
	cell's calculations, so we can blindly copy the calculated
	results, even if the =expression was adjusted wrt the cell
	addresses, as long as no error was tagged.

NOTE: For files with lot of cells having pseudo expressions and or
expressions which dont depend on other cells, for example filledexpr.csv,
the insert operation now is around 3.5-4 times faster.

	At same time the actual insert operation itself is now around
	10 times faster. Then there is the dependency list building,
	which takes same amount of time as before. So overall 3.5-4
	times faster.