Delete checks for fwdLinks to optimise
This is mostly similar to the optimisation done wrt insert in the
last commit.

However for cells which have fwdLinks i.e dependencies on other
cells, in case of insert, their old calc value may still be valid,
but with delete, it cant be guarenteed unless the full dependency
chain across cells (i.e cell-to-cell) is not validated to be equi
valent and values involved equal.

This brings similar performance gains as with insert (in last commit),
for spreadsheets having huge number of =expressions with no dependency
on other cells.

NOTE: Previous logic being replaced, also followed similar stratergy,
but was not extensive enough to filter out =expressions which are nops,
from cell dependency perspective. With syncd logic added, the more
extensive info is available to be used now.