RightAlign4: Moment of Initial truth/reality
Needed to fix a small oversight wrt col sliding range.

Also noticed the two issues which I know would/should occur if things
work out as expected. They are

1. As sData is not constrained/clipped when it is freshly picked
from its original cell, the original cell will contain the first
part of a long line, rather than the last part.

2. As a small content which doesnt fill the full cell width, is
currently left aligned by the cellstr logic, so we will have
situations where a small begining part of a long line will hang
on its own with spaces next to it, before the remaining part of
that long line appears on cells to the right of the cell at the
begining tip of a overflown text.

Added the global variable which maintains the current global
alignment state and inturn set it up for left alignment.

NOTE: Also in a previous commit I had renamed the

DATACOLSTART_OVERSCAN to be just DATACOL_OVERSCAN