use single matching entity from tree
irrespective of whether the token full matchs the entity in the tree,
or is the starting part of the only valid match in the tree.

instead of what the user had entered. This ensures that the single
matching entity of the tree gets returned to user, even if only a
partial part of it was entered by the user.