Allow tabbing thro non predefined parts of commands
Expanded some of the commands wrt their non predefined parts.

Added logic to skip through the non predefined parts and return,
them unchanged to the user.

Previously any part beyond the predefined part would have got
removed when the tab completion logic returns to the caller.

Now only the parts which shouldnt exist will be discarded, otherwise
the user controlled parts are left as is.

TODO: In the tree db, I have marked the user controlled parts as
being numeric or alpha numeric, later need to cross check the
user specified token matches this or not.