Handle perfect match tokens which are also prefixes
If such a token is a inbetween token with more tokens following
it, then chances are user wants to use the token, rather than its
siblings.

If such a token is the last token and the user has explicitly
added space to the end, then chances are user wants to use the token.

If not either of the above, then cycle through the other siblings,
which have a prefix which matchs this token.