Avoid additional chars in taber tree db definitions
As token parser is configured to treat only normal expected chars
to be part of the tokens, so avoid chars like _ or [ or ] or | in
the ANY char class suffixes in taber tree db pattern definitions.

NOTE: These suffixes to ANY char classes are only for user reference
and the program doesnt use it otherwise.