Allow cycling through all possible commands again
When the check to honar space if any at the end of input string
was added, there was a oversight in that I didnt check if the
string is empty or not, so it was raising a exception, thus
avoiding cycling through all possible commands if user doesnt
enter anything.

THis oversight has been fixed now.