Trap int to float overflow and related beyond
a very very long integer can overflow the float representation,
in which case any cells beyond the cell with the very very large
value, will be visible. But otherwise everything else will be
fine.

THis was a path from yesterdays cformat int to float conv. Have
trapped this path so that remaining cells can continue to behave
as configured/requested by user, and inturn this cell reverts to
the int value.

NOTE: Was to be expected as integers have arbitrary precision,
in python.