Account for single quotes in rsearch string argument
One needs to use single quotes around the string to search, if it
contains spaces in it. This is accounted for now.

NOTE: This is required because :rsearch will be extended to support
search and replace. In that case one requires to distinguish btw
the string to search, the replace keyword and the new string to use.