Allow logic to fly more freely by opening up recursion limit
Given that most of the initial/basic optimisation thoughts that
came to mind, around the different aspects of the spreadsheet logic,
has been implemented and checked (to some extent) to do its job to
a decent level. So extend the recursion depth allowed, so that the
program can fly more freely.

This inturn as allowed the program to perform atleast 4 times, if
not more times, faster.

NOTE: Constrain and test out the complex paths, and then open up
the easy path, so that the complex paths can be tested to some/
more extent than otherwise.