csvload:getCellAddrsIncRanges: states for celladdr ranges
added initial logic to handle some parts of the cell addr ranges
state logic.

If the state is setup as required else where, then either append
to caList as a single cell address or append to the caList as
part of a cell address range (by appending to last item already
in caList).

TODO: I require to advance the cellAddrRange state on seeing :

Also removed the old celladdrs_incranges_fromre related code, which
I had copied around for reference.