csvload:CGetCAIncR: iCARangeState CA afterCA without Range
The CARange state logic wasnt handling the case, when a CA follows
another CA and is also part of a CARange.

A CA has to be treated as standalone irrespective of whether it is

	the first CA or
	a CA following another CA without a range marker inbetween or
	a CA following a CARange.

And in all these 3 cases allow for the possibility of this CA being
the 1st part of a CA Range. Previously this was not handled for the
2nd possiblity mentioned above in the list of 3 possibilities.

This is handled now.