Fix oversight wrt lineEnd newLine check
When moving from python to c based logic for load line, made a mistake.

In it when checking for presence of newLine char at end of the string,
I was wrongly checking for char at lineLen instead of lineLen-1, this
inturn was leading to the last column of the row having newLine char
in it and inturn the rowAddr column of screen getting messed up, when
one scrolls to the extreme right of the spreadsheet, which inturn makes
the last column visible on the screen.

This has been fixed now.

Also changed the debug prints to use fprintf and inturn write to stderr,
so that using prg 2> /tmp/logfile, I can see what the program is doing.

Also have embedded the string being printed inside [] brackets, so that
one can know if there is newline in them or not easily.